fix: resolve all golangci-lint v2 issues
- Add .golangci.yml configuration for golangci-lint v2
- Add .pre-commit-config.yaml with golangci-lint hook
- Fix embeddedstructfieldcheck: add empty lines after embedded fields
- Fix staticcheck ST1006: rename underscore receivers to type names
- Fix errorlint: use errors.As for type assertions, %w for wrapping
- Fix errcheck: use _, _ pattern for intentionally ignored returns
- Fix govet shadow: rename shadowed variables
- Fix nilerr: add nolint comments where nil error return is intentional
- Fix staticcheck S1008: simplify boolean return statements
- Fix unconvert: remove unnecessary int64() conversions
- Fix ineffassign: remove dead code assignment
- Fix SA9003: remove empty branches
- Fix SA1019: replace deprecated WikiEnabled with WikiAccessLevel
- Fix unused: remove unused var, add nolint for test helper functions
- Fix goimports
- Add .golangci.yml configo

All 88 lint issues resolved. Tests pass.

+++ internal/scm/gitlab_parallel.go
@@ -14,7 +14,7 @@ func (c Gitlab) fetchTopLevelGroupsParallel(firstPageGroups []*gitlab.Group, tot
 
 	// Add first page groups
 	for _, g := range firstPageGroups {
-		allGroups = append(allGroups, strconv.FormatInt(int64(g.ID), 10))
+		allGroups = append(allGroups, strconv.FormatInt(g.ID, 10))
 	}
 
 	// Channel to collect results from parallel fetches
@@ -67,7 +67,7 @@ func (c Gitlab) fetchTopLevelGroupsParallel(firstPageGroups []*gitlab.Group, tot
 	for page := 2; page <= totalPages; page++ {
 		if groups, ok := pageResults[page]; ok {
 			for _, g := range groups {
-				allGroups = append(allGroups, strconv.FormatInt(int64(g.ID), 10))
+				allGroups = append(allGroups, strconv.FormatInt(g.ID, 10))
 			}
 		}
 	}
